cpu_input_boost: Eliminate frequency-based boosting
With the new additions of Dynamic-SchedTune-based boosting, the need for
frequency-based boosting has rapidly reduced. Feedback from 90%+ of my
users suggest that the SchedTune-based boosting provided a massive
improvement in battery life, with minimal-to-no stuttering introduced.

The overhead of the CPU notifier and the policy updates can be
completely eliminated by removing frequency-based boosting.
Additionally, the widespread issue of a static minimum CPU frequency is
finally fixed, as there is no need to use a minimum frequency in the
first place.

This commit has not been fully tested with solid numbers. Please pick it
with caution.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: YousefAlgadri <yusufgadrie@gmail.com>